Add tool definitions for Cloud Functions, GKE, Cloud Logging, Cloud Run, Cloud SQL, Cloud Storage, and Organization Policy Management
- Implemented definitions for Cloud Functions including listing and retrieving function details.
- Added GKE tools for listing clusters and retrieving cluster details.
- Created Cloud Logging tools for querying logs and listing log entries.
- Introduced Cloud Run tools for listing services and retrieving service details.
- Added definitions for Cloud SQL tools to list instances and get instance details.
- Implemented Cloud Storage tools for listing buckets, listing objects, and retrieving object metadata.
- Created Organization Policy Management tools for listing, getting, setting, and deleting organization policies.
- Updated index file to include new tool definitions and adjusted tests to accommodate changes.
